tarmac/new_pattern24/data-platform-devops-iac/terraform/fivetran/lambda/requirements.txt
boto3==1.28.66
botocore==1.31.66
dnspython
